# perf: Optimize spellcheck rule