This PR include refactoring the MatchedResult type and improvements to keyword matching and span handling:

- Refactored the MatchedResult type to use a tuple of the keyword length and a list of matched character indices instead of the `Span` struct.
- Updated the Node implementation to reflect the new MatchedResult type, simplifying the handling of spans.

run the `spellcheck` benchmark on my local pc:

```
spellcheck_50           time:   [4.2065 µs 4.2151 µs 4.2248 µs]
                        change: [-13.812% -13.537% -13.235%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

spellcheck_100          time:   [7.4984 µs 7.5490 µs 7.6227 µs]
                        change: [-14.877% -14.285% -13.617%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  7 (7.00%) high mild
  6 (6.00%) high severe

spellcheck_400          time:   [23.808 µs 23.841 µs 23.869 µs]
                        change: [-13.126% -12.951% -12.771%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)

```