@huacnlee 2025-03-23
我的想法是 struct Span 可以不用改，结构体和 Tuple 差异应该没那么绝大，相反 Span 对应其他有很多需要诊断信息的场景，理解起来会更清晰。

其他的可以合并。